Move timeout middleware into its own package
Relocate the ToolTimeout middleware from pkg/middleware to a dedicated
pkg/middleware/timeout package, mirroring the toolfilter package. This
keeps each middleware self-contained so future additions don't crowd a
single package.
