+++ pkg/middleware/timeout/timeout.go
@@ -1,4 +1,4 @@
-package middleware
+package timeout
 
 import (
 	"context"
