+++ pkg/middleware/timeout/timeout_test.go
@@ -1,4 +1,4 @@
-package middleware
+package timeout
 
 import (
 	"context"
